requirements.txt
allure-pytest==2.8.5
allure-python-commons==2.8.5
pytest==5.1.3
pytest-rerunfailures==7.0
pytest-runner==5.1
requests==2.22.0
selenium==3.141.0
webdriver-manager==1.8.2

